Moving math off into util/God/* classes.  Much cleaning of SWP.

+++ src/main/java/team3176/robot/subsystems/drivetrain/SwervePod2022Backup.java
@@ -49,7 +49,7 @@ public class SwervePod2022Backup {
     double spinEncoderPosition;
 
     private int id;
-    private int kEncoderOffset; 
+    private double kEncoderOffset; 
     private double kSpinEncoderUnitsPerRevolution;
     private double kDriveEncoderUnitsPerRevolution;
     private int off = 0;
